2-int_index.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,5 @@
 #include "function_pointers.h"
-
+#include <stdio.h>
 /**
  * array_iterater - ...
  * * @array: ...
@@ -9,7 +9,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int));
 {
-	unsigned int i = 0;
+	size_t i = 0;
 
 	if (array != NULL && action != NULL && size > 0)
 	{
